working middleware tested manually

+++ src/routers/project/projects.py
@@ -1,53 +1,36 @@
-from typing import Annotated
-from fastapi import APIRouter, HTTPException, Depends
+from fastapi import APIRouter, HTTPException, Depends, Request
 from starlette import status
 
 from sqlalchemy.orm import Session
 from src.project_handler_factory import createHandler
 from src.routers.project.schemas import NewProject, UpdateProject, Project
 from src.dependecies import get_db
-from jose import JWTError, jwt
-from fastapi.security import OAuth2PasswordBearer
-from src.services.auth_utils import SECRET_KEY, ALGORITHM, get_user
 
 project_router = APIRouter()
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")
-
-async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-    except JWTError:
-        raise credentials_exception
-    db = list(get_db())[0]
-    user = get_user(db=db, username=username)
-    if user is None:
-        raise credentials_exception
-    return user.username
-
 
 @project_router.get("/projects", response_model=list[Project])
-async def get_all_projects(db: Session = Depends(get_db),
-                           user_calling: str = Depends(get_current_user),
+async def get_all_projects(request: Request,
+                           db: Session = Depends(get_db),
                            project_handler: object = Depends(createHandler)):
-    try:    
-        return project_handler.get_all(db)
+    owned = request.headers["owned"]
+    participating = request.headers["participating"]
+    all = [owned, participating]
+    accessible = []
+    for li in [owned, participating]:
+        if len(li.rstrip()) != 0:
+            accessible = accessible + li.split(" ")
+    try:
+        return project_handler.get_all(db, accessible)
     except HTTPException as ex:
         raise ex
 
 
 @project_router.post("/projects")
-async def make_new_project(new_project: NewProject,
+async def make_new_project(request: Request,
+                           new_project: NewProject,
                            db: Session = Depends(get_db),
-                           user_calling: str = Depends(get_current_user),
                            project_handler: object = Depends(createHandler)):
+    user_calling = request.headers["username"]
     try:
         return project_handler.create(
             name=new_project.name,
@@ -59,10 +42,18 @@ async def make_new_project(new_project: NewProject,
         raise ex
 
 @project_router.get("/project/{project_id}/info", response_model=Project)
-async def get_project_details(project_id: int,
+async def get_project_details(request: Request,
+                              project_id: int,
                               db: Session = Depends(get_db),
-                              user_calling: str = Depends(get_current_user),
                               project_handler: object = Depends(createHandler)):
+    owned = request.headers["owned"]
+    print(owned.split(" "))
+    participating = request.headers["participating"]
+    print(participating.split(" "))
+    # check appropriate privileges
+    if str(project_id) not in owned.split(" ") and str(project_id) not in participating.split(" "):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="You don't have access to this project.")
     try:
         return project_handler.get(project_id, db)
     except HTTPException as ex:
@@ -70,16 +61,23 @@ async def get_project_details(project_id: int,
 
 
 @project_router.put("/project/{project_id}/info", response_model=Project)
-async def update_project_details(project_id: int,
+async def update_project_details(request: Request,
+                                 project_id: int,
                                  new_info: UpdateProject,
                                  db: Session = Depends(get_db),
-                                 user_calling: str = Depends(get_current_user),
                                  project_handler: object = Depends(createHandler)):
     if new_info.model_fields_set == set():
         raise HTTPException(
             status_code=400,
             detail="No project properties were specified in the request body"
         )
+    owned = request.headers["owned"]
+    participating = request.headers["participating"]
+    # check appropriate privileges
+    if str(project_id) not in owned.split(" ") and str(project_id) not in participating.split(" "):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="You don't have access to this project.")
+    user_calling = request.headers["username"]
     for_update = new_info.model_dump(exclude_unset=True)
     for_update.update({"updated_by":  user_calling})
     try:
@@ -89,10 +87,16 @@ async def update_project_details(project_id: int,
 
 
 @project_router.delete("/project/{project_id}")
-async def delete_project(project_id: int,
+async def delete_project(request: Request,
+                         project_id: int,
                          db: Session = Depends(get_db),
-                         user_calling: str = Depends(get_current_user),
                          project_handler: object = Depends(createHandler)):
+    owned = request.headers["owned"]
+    if str(project_id) not in owned:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Only project owner can delete it"
+        )        
     try:
         project_handler.delete(project_id, db)
         return status.HTTP_204_NO_CONTENT
@@ -101,11 +105,17 @@ async def delete_project(project_id: int,
     
 
 @project_router.post("/project/{project_id}/invite")
-async def add_collaborator(project_id: int,
+async def add_collaborator(request: Request,
+                           project_id: int,
                            user: str,
                            db: Session = Depends(get_db),
-                           user_calling: str = Depends(get_current_user),
                            project_handler: object = Depends(createHandler)):
+    owned = request.headers["owned"]
+    if str(project_id) not in owned:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Only project owner can invite participants"
+        )
     try:
         project_handler.grant_access(project_id, user, db)
     except HTTPException as ex:
